fix row accessibility issue add tests for isHeader
